poc-1-core-loop/src/requirements.txt
google-adk
python-dotenv # Used by main.py to load optional .env file